Implement new initiative selection
Removed the screen for selecting which PCs are in combat and their initiatives. These can now be determined using the checkbox next to the PCs name and the initiative spinner, respectively, on the pre combat view. Loading an encounter now only starts the music for that encounter without changing views until the encounter is actually started.